[BACKPORT] lib/mpi: fix bugs spotted with clang 14.0.1
Clang 14.0.1 reported a couple of bugs in lib/mpi code, this commit backports the necessary patches and fix them

This is a backport of the following commits:

[PATCH 1/2] lib/mpi: avoid assembler warning
Arnd Bergmann Fri, 26 Feb 2016 04:48:00 -0800

A wrapper around the umull assembly instruction might reuse
the input register as an output, which is undefined on
some ARM machines, as pointed out by this assembler warning:

  CC      lib/mpi/generic_mpih-mul1.o
/tmp/ccxJuxIy.s: Assembler messages:
/tmp/ccxJuxIy.s:53: rdhi, rdlo and rm must all be different
  CC      lib/mpi/generic_mpih-mul2.o
/tmp/ccI0scAD.s: Assembler messages:
/tmp/ccI0scAD.s:53: rdhi, rdlo and rm must all be different
  CC      lib/mpi/generic_mpih-mul3.o
/tmp/ccMvVQcp.s: Assembler messages:
/tmp/ccMvVQcp.s:53: rdhi, rdlo and rm must all be different

This changes the constraints to force different registers to
be used as output.

Signed-off-by: Arnd Bergmann <a...@arndb.de>
---

[PATCH 2/2] lib/mpi: use "static inline" instead of "extern inline"
Arnd Bergmann Fri, 26 Feb 2016 04:48:07 -0800

When we use CONFIG_PROFILE_ALL_BRANCHES, every 'if()' introduces
a static variable, but that is not allowed in 'extern inline'
functions:

mpi-inline.h:116:204: warning: '______f' is static but declared in inline
function 'mpihelp_sub' which is not static
mpi-inline.h:113:184: warning: '______f' is static but declared in inline
function 'mpihelp_sub' which is not static
mpi-inline.h:70:184: warning: '______f' is static but declared in inline
function 'mpihelp_add' which is not static
mpi-inline.h:56:204: warning: '______f' is static but declared in inline
function 'mpihelp_add_1' which is not static

This changes the MPI code to use 'static inline' instead, to get
rid of hundreds of warnings.

Signed-off-by: Arnd Bergmann <a...@arndb.de>
---

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>